Added SBC and ADC indirect support
This was a triumph, I'm making a note here huge success. It's hard to overstate my satisfaction. Aperture Science, we do what we must because we can. For the good of all of us except the ones who are dead. But there's no use crying over every mistake we just keep on trying 'till we run out of cake cause there's science to be done and experiments to be run on the people who are still alive.